backend/requirements_phase7.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-socketio==5.10.0
aiofiles==23.2.1
psutil==5.9.6
requests==2.31.0
sqlalchemy==2.0.23
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
opencv-python==4.8.1.78
librosa==0.10.1
soundfile==0.12.1
SpeechRecognition==3.10.0
pydub==0.25.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
O365==2.0.26
aiohttp==3.9.1
openai==1.3.7
pytesseract==0.3.10
pyaudio==0.2.11
pywin32 